This PR contains the following performance improvements:

* **Prevent eager creation of anyhow::Error during serialization:**
This change improved throughput by 25% for the throughput/parallel benchmark on my machine.

* **Read first from old invocation status table and then from the new one:**
This change did not result in a measurable performance improvement for throughput/parallel on my machine. Theoretically, it should be cheaper since we currently never write to the new invocation status table.

* **Batch apply partition processor commands before committing:**
During benchmarking we discovered that committing txn to RocksDB took up most of the CPU time. Before, every partition processor command resulted into a RocksDB transaction commit. Now, we are batching the available commands to create larger transaction that get committed less frequently. This change helped improve the throughput for the parallel benchmark by 40%. Batch applying the partition processor commands comes potentially at the cost of increased latency. To mitigate the negative effects on latency, the batching size is configurable with a default of `4`. 

@AhmedSoliman it would be great if you could run this PR with the wrk test scenario to validate that it does not affect latency too badly. If it does, then we might want to choose a different default value for the max batching size.