@AhmedSoliman 2024-08-21
Running the main branch without my other changes. 
Before your PR:
```console
❯ wrk -t 8 -c 200 --latency -d3m  -s ~/workspace/get.lua http://127.0.0.1:8080
Running 3m test @ http://127.0.0.1:8080
  8 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    32.69ms   31.84ms 781.46ms   98.73%
    Req/Sec   826.07     96.46     1.33k    82.57%
  Latency Distribution
     50%   30.46ms
     75%   33.65ms
     90%   37.05ms
     99%   73.35ms
  1177712 requests in 3.00m, 313.36MB read
Requests/sec:   6540.39
Transfer/sec:      1.74MB
```
After:
```console
❯ wrk -t 8 -c 200 --latency -d3m  -s ~/workspace/get.lua http://127.0.0.1:8080
Running 3m test @ http://127.0.0.1:8080
  8 threads and 200 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    38.49ms   82.42ms   1.61s    98.90%
    Req/Sec   796.85     87.34     1.21k    81.89%
  Latency Distribution
     50%   31.32ms
     75%   34.37ms
     90%   37.90ms
     99%  143.66ms
  1131366 requests in 3.00m, 301.03MB read
Requests/sec:   6282.92
Transfer/sec:      1.67MB
```