@AhmedSoliman 2024-08-21
The more important impact here is that this was about 13% less CPU after your changes, so it's definitely a net improvement despite the max/tail latency regression. @tillrohrmann did you consider making the max_batching live-configurable?